CLI-1043 Move the stage module under src/core/commands
