+++ CLAUDE.md
@@ -38,6 +38,7 @@ Each command lives in `src/commands/`. The command tree is built by `createComma
 
 To add a new command: add it to `createCommandTree()` in `src/commands/command-tree.ts` and implement the logic in a new folder under `src/commands/`.
 Please declare commands using the type defined in `src/commands/sonar-command.ts`.
+Lifecycle stages (`Stage`) live in `src/core/commands/stage.ts` and are re-exported from `sonar-command.ts`.
 By default, new commands should register a `authenticatedAction()`; the handler's first argument is a `CommandAuthenticatedInvocationContext` (`auth`, plus stage accessors). `anonymousAction()` receives a `CommandInvocationContext` parent. Both live in `src/commands/command-invocation-context.ts`. Only technical commands use `anonymousAction()`. On the context, `isAlphaEligible()` / `isBetaEligible()` answer whether **this execution** should be treated as alpha/beta — not merely whether `.stage()` names that lifecycle: `isAlphaEligible()` requires Alpha stage and `SONARQUBE_CLI_ALPHA`; `isBetaEligible()` is true for Open Beta, or for Private Beta when the user is entitled for the command's LaunchDarkly flag. SonarCommand's own `isAlpha` / `isBeta` / `isDeprecated` getters remain stage-name checks used for tree visibility and help tags. Handlers record telemetry with `ctx.recordTelemetry(new TelemetryFact(name, payload))` — timestamp is set on construction and can be overridden; `payload` is the bare domain data only (opaque to core). Facts live on that invocation context (`ctx.telemetryFacts()`), not a tree-owned buffer. `postAction` reads them from the action command's context, resolves the agent session id, then `commitTelemetryFacts` (`src/commands/telemetry-facts.ts`) emits handler facts concatenated with a `CliCommandExecuted` fact (`buildCommandExecutedFact` in `src/commands/command-executed-telemetry.ts`) through generic `emitTelemetryEvent` and schedules the flush.
 Mark lifecycle with `.stage(Stage.Stable)` (default), `.stage(Stage.Alpha)`, `.stage(Stage.Beta())` (Open Beta), `.stage(Stage.Beta('cli.beta.flag-key'))` (Private Beta, gated by LaunchDarkly), or `.stage(Stage.Deprecated({ sinceVersion, replacementCommand }))`. Visibility is decided via `CliRuntime` (`isAlphaEnabled` from `SONARQUBE_CLI_ALPHA`, `isPrivateBetaEnabled` from LaunchDarkly): `.stage()` keeps or detaches the command after Commander’s eager `.command()` attach. Deprecated commands stay visible and warn on every invocation (`replacementCommand` is required; pass `null` when there is none). `createCommandTree` is async: it probes for Private Beta flag keys, then only when keys exist calls optional `loadPrivateBetaContext(flagKeys)` (wired in `src/index.ts` to resolve auth + `resolvePrivateBetaFlags`, 2-hour cache under `~/.sonar/sonarqube-cli/launch-darkly/`) and rebuilds the tree so Private Beta commands register only when the flag is `true`. With no keys (or no loader), LaunchDarkly is never contacted; docs call `await createCommandTree()` without a loader so Private Beta stays omitted. Set `SONARQUBE_CLI_LAUNCHDARKLY_ENV=dev` to use the LaunchDarkly `dev` environment (default `production`). Open Beta and Deprecated appear in public docs; Private Beta and Alpha do not.
 Contributors stage options with `addOption(new SonarOption('--flag', 'Help').stage(Stage.Alpha))` — `.option().stage()` cannot work because `.option()` returns the command. Use `.stage(Stage.Beta())` for Open Beta, `.stage(Stage.Beta('cli.beta.flag-key'))` for Private Beta, and `.stage(Stage.Deprecated({ sinceVersion, replacementCommand }))` for Deprecated. Required/mandatory options cannot be staged as Alpha or Beta; they can be Deprecated. When the caller is not entitled, Alpha and Private Beta options are omitted from help and treated as unknown. Docs include Open Beta and Deprecated options (with badges) and omit Alpha and Private Beta options. Invoking a deprecated command or passing a deprecated option prints a warning every time (`Use '<command>' instead.` or `There is no replacement.`).
