+++ src/core/commands/stage.ts
